clear button

+++ gothonweb/app2.py
@@ -87,10 +87,12 @@ class Score(object):
 
 class Clear(object):
     def GET(self):
-        os.remove("scores.json")
-        del lst[:]
-        os.remove("scores2.json")
-        del lst2[:]
+		if os.path.isfile("scores.json"):        
+			os.remove("scores.json")
+			del lst[:]
+		if os.path.isfile("scores2.json"):
+			os.remove("scores2.json")
+			del lst2[:]
         web.seeother("/score")
 
 class GameEngine(object):
